ICompressor and DummyCompressor moved to vmfcore
