+++ modules/vmfcore/src/icompressor.cpp
